# openapi.json
{
  "openapi": "3.0.0",
  "paths": {
...
    "/cloud-api/auth/signup": {
      "post": {
        "description": "Creates a new user account and initiates the onboarding flow. Returns an access token and sets a refresh token in an httpOnly cookie.",
        "operationId": "AuthController_signup",
        "parameters": [
          {
            "name": "user-agent",
            "required": true,
            "in": "header",
            "schema": {
              "type": "string"
            }
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/SignupDto"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "User created successfully. Returns onboarding status and access token."
          },
          "400": {
            "description": "Invalid input data or validation error."
          },
          "409": {
            "description": "User with this email already exists."
          }
        },
        "summary": "User signup",
        "tags": [
          "Auth"
        ]
      }
    },
    "/cloud-api/auth/token": {
      "get": {
        "description": "Recovers the session by reading the refresh token from the httpOnly cookie and returns a new access token. Does not rotate the refresh token.",
        "operationId": "AuthController_getToken",
        "parameters": [],
        "responses": {
          "200": {
            "description": "Session recovered successfully. Returns new access token.",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "properties": {
                    "accessToken": {
                      "type": "string",
                      "description": "JWT access token"
                    },
                    "expiresIn": {
                      "type": "number",
                      "description": "Token expiry in seconds"
                    }
                  }
                }
              }
            }
          },
          "401": {
            "description": "Invalid or expired refresh token."
          }
        },
        "summary": "Recover session token",
        "tags": [
          "Auth"
        ]
      }
    },
    "/cloud-api/auth/login": {
      "post": {
        "description": "Authenticates the user with email and password. Returns an access token and sets a refresh token in an httpOnly cookie.",
        "operationId": "AuthController_login",
        "parameters": [
          {
            "name": "user-agent",
            "required": true,
            "in": "header",
            "schema": {
              "type": "string"
            }
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/LoginDto"
              }
            }
          }
        },
        "responses": {
          "201": {
            "description": "Login successful. Returns access token and user information."
          },
          "400": {
            "description": "Invalid input data or validation error."
          },
          "401": {
            "description": "Invalid credentials."
          },
          "404": {
            "description": "User not found."
          }
        },
        "summary": "User login",
        "tags": [
          "Auth"
        ]
      }
    },
    "/cloud-api/auth/refresh": {
      "post": {
        "description": "Generates a new access token and rotates the refresh token for enhanced security. Reads refresh token from httpOnly cookie and updates it with the new rotated token.",
        "operationId": "AuthController_refreshToken",
        "parameters": [],
        "responses": {
          "201": {
            "description": "Token refreshed successfully. Returns new access token.",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "properties": {
                    "accessToken": {
                      "type": "string",
                      "description": "New JWT access token"
                    },
                    "expiresIn": {
                      "type": "number",
                      "description": "Token expiry in seconds"
                    }
                  }
                }
              }
            }
          },
          "401": {
            "description": "Invalid or expired refresh token."
          }
        },
        "summary": "Refresh access token",
        "tags": [
          "Auth"
        ]
      }
    },
    "/cloud-api/auth/logout": {
      "post": {
        "description": "Invalidates the current session and clears the refresh token cookie. Only logs out from the current device.",
        "operationId": "AuthController_logout",
        "parameters": [],
        "responses": {
          "201": {
            "description": "Successfully logged out.",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "properties": {
                    "message": {
                      "type": "string",
                      "example": "Successfully logged out"
                    }
                  }
                }
              }
            }
          },
          "401": {
            "description": "Unauthorized. Invalid or missing access token."
          }
        },
        "security": [
          {
            "bearer": []
          }
        ],
        "summary": "Logout from current device",
        "tags": [
          "Auth"
        ]
      }
    },
    "/cloud-api/auth/logout-all": {
      "post": {
        "description": "Invalidates all active sessions for the current user across all devices and clears the refresh token cookie.",
        "operationId": "AuthController_logoutAll",
        "parameters": [],
        "responses": {
          "201": {
            "description": "Successfully logged out from all devices.",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "properties": {
                    "message": {
                      "type": "string",
                      "example": "Successfully logged out from 3 device(s)"
                    }
                  }
                }
              }
            }
          },
          "401": {
            "description": "Unauthorized. Invalid or missing access token."
          }
        },
        "security": [
          {
            "bearer": []
          }
        ],
        "summary": "Logout from all devices",
        "tags": [
          "Auth"
        ]
      }
    },
...
    "/cloud-api/auth/forgot-password": {
      "post": {
        "description": "Sends a password reset OTP to the provided email address. Always returns success to prevent email enumeration.",
        "operationId": "AuthController_forgotPassword",
        "parameters": [],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/ForgotPasswordDto"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Password reset email sent (if account exists).",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "properties": {
                    "success": {
                      "type": "boolean",
                      "example": true
                    },
                    "message": {
                      "type": "string",
                      "example": "If an account with that email exists, a password reset code has been sent."
                    }
                  }
                }
              }
            }
          },
          "400": {
            "description": "Invalid input data."
          }
        },
        "summary": "Request password reset",
        "tags": [
          "Auth"
        ]
      }
    },
    "/cloud-api/auth/verify-reset-otp": {
      "post": {
        "description": "Validates the OTP sent to the user email and returns a reset token for setting a new password.",
        "operationId": "AuthController_verifyResetOtp",
        "parameters": [],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/VerifyResetOtpDto"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "OTP verified successfully. Returns reset token.",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "properties": {
                    "resetToken": {
                      "type": "string",
                      "description": "Token to use for resetting the password"
                    }
                  }
                }
              }
            }
          },
          "400": {
            "description": "No reset request found or OTP expired."
          },
          "401": {
            "description": "Invalid OTP."
          }
        },
        "summary": "Verify password reset OTP",
        "tags": [
          "Auth"
        ]
      }
    },
    "/cloud-api/auth/reset-password": {
      "post": {
        "description": "Sets a new password using the reset token received after OTP verification. Invalidates all active sessions.",
        "operationId": "AuthController_resetPassword",
        "parameters": [],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/ResetPasswordDto"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Password reset successfully.",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "properties": {
                    "success": {
                      "type": "boolean",
                      "example": true
                    },
                    "message": {
                      "type": "string",
                      "example": "Password has been reset successfully. Please login with your new password."
                    }
                  }
                }
              }
            }
          },
          "400": {
            "description": "Invalid or expired reset token."
          }
        },
        "summary": "Reset password",
        "tags": [
          "Auth"
        ]
      }
    },
    "/cloud-api/auth/password/change": {
      "post": {
        "description": "Change the authenticated user's password. Requires current password verification.",
        "operationId": "AuthController_changePassword",
        "parameters": [],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/ChangePasswordDto"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Password changed successfully",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "properties": {
                    "message": {
                      "type": "string",
                      "example": "Password changed successfully"
                    }
                  }
                }
              }
            }
          },
          "400": {
            "description": "Invalid current password or validation error"
          },
          "401": {
            "description": "Unauthorized - Invalid or missing authentication"
          }
        },
        "security": [
          {
            "bearer": []
          }
        ],
        "summary": "Change password",
        "tags": [
          "Auth"
        ]
      }
    },
    "/cloud-api/auth/sessions": {
      "get": {
        "description": "Get all active sessions for the authenticated user across all devices.",
        "operationId": "AuthController_getSessions",
        "parameters": [],
        "responses": {
          "200": {
            "description": "Active sessions retrieved successfully",
            "content": {
              "application/json": {
                "schema": {
                  "type": "array",
                  "items": {
                    "$ref": "#/components/schemas/SessionResponseDto"
                  }
                }
              }
            }
          },
          "401": {
            "description": "Unauthorized - Invalid or missing authentication"
          }
        },
        "security": [
          {
            "bearer": []
          }
        ],
        "summary": "List active sessions",
        "tags": [
          "Auth"
        ]
      }
    },
    "/cloud-api/auth/sessions/{id}": {
      "delete": {
        "description": "Invalidate a specific session by ID. Cannot revoke the current session.",
        "operationId": "AuthController_revokeSession",
        "parameters": [
          {
            "name": "id",
            "required": true,
            "in": "path",
            "description": "Session ID to revoke",
            "schema": {
              "example": "550e8400-e29b-41d4-a716-446655440000",
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Session revoked successfully",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "properties": {
                    "message": {
                      "type": "string",
                      "example": "Session revoked successfully"
                    }
                  }
                }
              }
            }
          },
          "400": {
            "description": "Cannot revoke current session"
          },
          "401": {
            "description": "Unauthorized - Invalid or missing authentication"
          },
          "404": {
            "description": "Session not found"
          }
        },
        "security": [
          {
            "bearer": []
          }
        ],
        "summary": "Revoke a specific session",
        "tags": [
          "Auth"
        ]
      }
    },
...
  },
  "info": {
    "title": "Vritti Cloud API",
    "description": "Internal API for Vritti SaaS Platform",
    "version": "1.0.0",
    "contact": {}
  },
...
  "servers": [
    {
      "url": "https://local.vrittiai.com:3000",
      "description": "Local Development"
    }
  ],
  "components": {
    "securitySchemes": {
      "bearer": {
        "scheme": "bearer",
        "bearerFormat": "JWT",
        "type": "http",
        "description": "Enter your JWT access token"
      }
    },
    "schemas": {
...
      "SignupDto": {
        "type": "object",
        "properties": {
          "email": {
            "type": "string",
            "description": "User email address",
            "example": "user@example.com",
            "format": "email"
          },
          "password": {
            "type": "string",
            "description": "Password (min 8 chars, must include uppercase, lowercase, number, and special character)",
            "example": "SecureP@ss123",
            "minLength": 8
          },
          "firstName": {
            "type": "string",
            "description": "User first name",
            "example": "John"
          },
          "lastName": {
            "type": "string",
            "description": "User last name",
            "example": "Doe"
          }
        },
        "required": [
          "email",
          "password"
        ]
      },
      "LoginDto": {
        "type": "object",
        "properties": {
          "email": {
            "type": "string",
            "description": "User email address",
            "example": "user@example.com",
            "format": "email"
          },
          "password": {
            "type": "string",
            "description": "User password",
            "example": "SecureP@ss123",
            "minLength": 8
          }
        },
        "required": [
          "email",
          "password"
        ]
      },
...
      "ForgotPasswordDto": {
        "type": "object",
        "properties": {
          "email": {
            "type": "string",
            "description": "User email address",
            "example": "user@example.com",
            "format": "email"
          }
        },
        "required": [
          "email"
        ]
      },
      "VerifyResetOtpDto": {
        "type": "object",
        "properties": {
          "email": {
            "type": "string",
            "description": "User email address",
            "example": "user@example.com",
            "format": "email"
          },
          "otp": {
            "type": "string",
            "description": "6 digit OTP sent to the user email",
            "example": "123456"
          }
        },
        "required": [
          "email",
          "otp"
        ]
      },
      "ResetPasswordDto": {
        "type": "object",
        "properties": {
          "resetToken": {
            "type": "string",
            "description": "Reset token received after OTP verification"
          },
          "newPassword": {
            "type": "string",
            "description": "New password (min 8 chars, must include uppercase, lowercase, number, and special character)",
            "example": "NewPassword123!",
            "minLength": 8
          }
        },
        "required": [
          "resetToken",
          "newPassword"
        ]
      },
      "ChangePasswordDto": {
        "type": "object",
        "properties": {
          "currentPassword": {
            "type": "string",
            "description": "Current password for verification",
            "example": "MyOldPassword123!"
          },
          "newPassword": {
            "type": "string",
            "description": "New password (min 8 chars, must contain uppercase, lowercase, number, and special character)",
            "example": "MyNewPassword456!"
          }
        },
        "required": [
          "currentPassword",
          "newPassword"
        ]
      },
      "SessionResponseDto": {
        "type": "object",
        "properties": {
          "sessionId": {
            "type": "string",
            "description": "Unique session identifier",
            "example": "550e8400-e29b-41d4-a716-446655440000"
          },
          "device": {
            "type": "string",
            "description": "Device information parsed from user agent",
            "example": "Chrome on macOS"
          },
          "location": {
            "type": "string",
            "description": "Geographic location based on IP address",
            "example": "San Francisco, CA"
          },
          "ipAddress": {
            "type": "string",
            "description": "IP address of the session",
            "example": "192.168.1.1"
          },
          "lastActive": {
            "format": "date-time",
            "type": "string",
            "description": "Last activity timestamp",
            "example": "2026-02-04T12:30:00Z"
          },
          "isCurrent": {
            "type": "boolean",
            "description": "Whether this is the current session",
            "example": false
          }
        },
        "required": [
          "sessionId",
          "device",
          "location",
          "ipAddress",
          "lastActive",
          "isCurrent"
        ]
      },
...
    }
  }
}